Initial version of jnTAC
This is the initial version of jnTAC. I am at the very beginning. The game is divided into three main classes:
- Player: Implementing players.
- Mumble: Implementing mumbles.
- Card:   Implementing cards.
I am unhappy with the card class. Since it is a superclass for several other subclasses I am not able to access instance fields or methods of the subclasses. A parent class cannot access the fields or methods of a child class, since it has no knowledge of what the child class has implemented. So next i want to try to correct this issue.