Implemented interface Card.java.
Card is now an interface. The classes MasterCard and NumberedCard implement the interface. I will define all available cards in the game inside those classes. The method playSelectedCard(...) is defined in all subclasses which implement the Card interface. This allows us to use the cards independent of the class of the specific card. If you want to use a NumberedSpecialCard you just apply the playSelectedCard on the specified card and the appropriate playSelectedCard method from the same class (in this example NumberedSpecialCard) will be applied to the card object.

I implemented a first draft of the drawCard method in the Player.java class. Later this method will draw new cards when the player used all his cards.